#17 編集機能のhooks化
